update version, tweak more textures, remove death ray debug item, wind balls improved visuals, added more blocks to map colors, removed unused trap block, updated lever recipe
