+++ gradle.properties
@@ -16,6 +16,6 @@ btwaila_version=1.2.2-7.3_04
 uselessnumerical_version=1.0.2
 
 # Mod
-mod_version=0.8.0+7.3_04
+mod_version=1.0.0+7.3_04
 mod_group=teamport
 mod_name=better-with-aether
